Extract methods related to persisted preferences to module
This removes even more code from Spree::Base, and groups the few lines
of code that are needed for persisted preferences in one module.
