+++ core/app/models/spree/promotion_action.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
+require 'spree/preferences/persistable'
+
 module Spree
   # Base class for all types of promotion action.
   #
   # PromotionActions perform the necessary tasks when a promotion is activated
   # by an event and determined to be eligible.
   class PromotionAction < Spree::Base
-    serialize :preferences, Hash
-    after_initialize :initialize_preference_defaults
-
+    include Spree::Preferences::Persistable
     include Spree::SoftDeletable
 
     belongs_to :promotion, class_name: 'Spree::Promotion', inverse_of: :promotion_actions, optional: true
