+++ core/app/models/spree/promotion_rule.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
+require 'spree/preferences/persistable'
+
 module Spree
   # Base class for all promotion rules
   class PromotionRule < Spree::Base
-    serialize :preferences, Hash
-    after_initialize :initialize_preference_defaults
+    include Spree::Preferences::Persistable
 
     belongs_to :promotion, class_name: 'Spree::Promotion', inverse_of: :promotion_rules, optional: true
